backend/requirements.txt
fastapi==0.115.6
matplotlib==3.10.0
opencv-python==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.0
seaborn==0.13.2
setuptools==75.6.0
tqdm==4.67.1
ultralytics==8.3.59
uvicorn==0.34.0
timm==1.0.15
scikit-learn==1.6.0
itsdangerous==2.2.0
wandb==0.19.1
python-multipart==0.0.20
psycopg2-binary
pyjwt==2.10.1
bcrypt==4.2.1
httpx==0.28.1
psutil==5.9.8
apscheduler==3.9.1
numpy==1.24.3
gdown==5.2.0
coolname==2.2.0
huggingface_hub==0.30.2
transformers==4.51.3
rarfile==4.2
jenkspy==0.4.1
PytorchWildlife==1.2.3